refactor: refactor project structure

+++ backend/go_services/metadata/internal/domain/entities/genre.go
@@ -1,7 +1,7 @@
 package entities
 
 type Genre struct {
-	ID      uint8     `gorm:"primaryKey"`
+	ID      int64     `gorm:"primaryKey"`
 	Name    string    `gorm:"column:name;not null;unique;size:25;"`
 	Artists []*Artist `gorm:"many2many:artist_genres"`
 	Albums  []*Album  `gorm:"many2many:album_genres"`
